move to string.ts
