+++ packages/grafana-ui/src/utils/index.ts
@@ -2,5 +2,5 @@ export * from './processTimeSeries';
 export * from './valueFormats/valueFormats';
 export * from './colors';
 export * from './namedColorsPalette';
-export * from './stringUtils';
+export * from './string';
 export { getMappedValue } from './valueMappings';
