+++ packages/grafana-ui/src/utils/string.test.ts
