+++ packages/grafana-ui/src/utils/string.ts
